adding a screen in case of no events

+++ utils/helpers.js
@@ -2,7 +2,7 @@ import dayjs from "dayjs"
 
 function getEndTime(startTime, duration){
     let seperatedDuration = duration.split(" ")
-    const firstParam = parseInt(seperatedDuration[0])
+    const firstParam = parseFloat(seperatedDuration[0])
     
     let endTime = dayjs(startTime).add(firstParam, "h").format("hh:mm a")
